OCMockObject is not thread safe
Make OCMockObject thread safe.

Synchronize on the four mutable array instance variables during access to ensure their consistency.
Retain everything retrieved from the arrays if the retrieved objects ever get used outside of the synchronized chunks.

Retain captured invocation arguments since the arguments could otherwise be deallocated at any time.  The invocation can't outright retain its arguments since that would cause a retain cycle, so hold a copy of the invocation that has self removed, and retain the arguments on that.  This also causes a problem with partial mocks because they set up a forwarder for -retain. The forwarder would capture the call to -retain as an NSInvocation, OCMockObject now retains the invocation arguments, which calls -retain on the original object again, which hits the forwarder method again, and so on and so forth. Break the cycle by adding all the reference counting methods (-retainCount, -retain, -release, -autorelease) to OCPartialMockObject's selector black list.

Make OCMMacroState thread safe as well.

Having a global macro state doesn't work.  With locking it doesn't crash, but if two threads are recording at the same time then one of the macro states gets stepped on.  Change the global state to be a per-thread state.
Avoid a use-after-dealloc race by not trying to clear the global state in -dealloc and making the macros exception safe such that the +endXXXMacro calls always happen even if the captured invocation throws an exception.

Make OCObserverMockObject thread safe.

Synchronize on the two mutable array instance variables during access to ensure their consistency.

https://github.com/erikdoe/ocmock/issues/171